% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{add_transform}
\alias{add_transform}
\title{Add Transform to Model}
\usage{
add_transform(model, subject, transform)
}
\arguments{
\item{model}{An AlignmentModel object.}

\item{subject}{Subject ID for the new transform.}

\item{transform}{The transform operator.}
}
\value{
A new AlignmentModel with the added transform.
}
\description{
Creates a new model with an additional transform (does not modify in place).
}
