% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{aligned_data}
\alias{aligned_data}
\title{Get All Aligned Data as List}
\usage{
aligned_data(result)
}
\arguments{
\item{result}{An AlignmentResult object.}
}
\value{
Named list of aligned data.
}
\description{
Get All Aligned Data as List
}
