% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligned_resample_set.R
\name{AlignedResampleSet}
\alias{AlignedResampleSet}
\title{Create an AlignedResampleSet}
\usage{
AlignedResampleSet(
  splits,
  resampling_spec = list(),
  aggregation_contract = list(aggregate = c("metrics", "predictions"), forbid =
    c("raw_coordinates", "stack_shared_features")),
  metadata = list()
)
}
\arguments{
\item{splits}{Named list of retained split records.}

\item{resampling_spec}{Resampling description.}

\item{aggregation_contract}{Aggregation policy list.}

\item{metadata}{Extra metadata.}
}
\value{
An \link{AlignedResampleSet}.
}
\description{
Create an AlignedResampleSet
}
\seealso{
Other aligned_study:
\code{\link{AlignedBlock}()},
\code{\link{AlignedResampleSet-class}},
\code{\link{AlignedStudy}},
\code{\link{AlignedStudy-class}},
\code{\link{SharedFeatureSpace}()},
\code{\link{align_study}()},
\code{\link{aligned_matrix}()},
\code{\link{alignment_model}()},
\code{\link{analysis_lineage}()},
\code{\link{analysis_safety}()},
\code{\link{analysis_safety_record}()},
\code{\link{as_aligned_resample_set}()},
\code{\link{as_aligned_study}()},
\code{\link{as_alignment_data.AlignedStudy}()},
\code{\link{as_hyperdesign}()},
\code{\link{as_subject_matrices}()},
\code{\link{assert_analysis_safe}()},
\code{\link{assert_common_shared_space}()},
\code{\link{feature_data}()},
\code{\link{map_subjects}()},
\code{\link{observation_data}()},
\code{\link{orientation}},
\code{\link{shared_feature_space_from_model}()},
\code{\link{shared_space}()},
\code{\link{shared_space_id}()},
\code{\link{shared_spaces_compatible}()},
\code{\link{stack_aligned_resamples}()},
\code{\link{stack_subjects}()},
\code{\link{subject_data}()},
\code{\link{subject_ids}()}
}
\concept{aligned_study}
