% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_cone.R
\name{aligner_cone}
\alias{aligner_cone}
\title{CONE-Align}
\description{
Consensus Optimisation for Node Embedding Alignment via manifoldalign.
Iteratively alternates between Procrustes rotation of spectral
embeddings and Hungarian assignment of node correspondences.
}
\keyword{internal}
