% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes.R
\name{aligner_procrustes}
\alias{aligner_procrustes}
\title{Procrustes Alignment}
\description{
Procrustes alignment via manifoldalign. Finds orthogonal transformations
that minimize the sum of squared distances between subjects' data.
}
