% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_fmrigds.R
\name{alignment_data_from_gds}
\alias{alignment_data_from_gds}
\title{Create AlignmentData from gds_data}
\usage{
alignment_data_from_gds(gds_data, assay = "beta", ...)
}
\arguments{
\item{gds_data}{A gds_data object from fmrigds.}

\item{assay}{Name of assay to extract (default: "beta"). Extracted as a
per-subject matrix (samples x contrasts).}

\item{...}{Additional arguments passed to AlignmentData.}
}
\value{
An AlignmentData object.
}
\description{
Extract data from an fmrigds gds_data object into AlignmentData format.
}
\examples{
\dontrun{
# Requires fmrigds and a gds_data object:
# adat <- alignment_data_from_gds(gds, assay = "beta")
# res <- fit_alignment(adat, method = "procrustes")
}

}
