% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{as_aligned_matrix}
\alias{as_aligned_matrix}
\title{Convert Aligned Data to Matrix}
\usage{
as_aligned_matrix(result, by = c("subject", "observation"))
}
\arguments{
\item{result}{An AlignmentResult object.}

\item{by}{How to stack: "subject" (list element per subject) or
"observation" (concatenate columns).}
}
\value{
Matrix or list of matrices.
}
\description{
Stack all aligned subjects into a single matrix.
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(20), 4, 5),
  s2 = matrix(rnorm(20), 4, 5)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
mat <- as_aligned_matrix(res, by = "observation")
dim(mat)

}
