% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{compute_centroid}
\alias{compute_centroid}
\title{Compute Mean/Centroid Data}
\usage{
compute_centroid(data)
}
\arguments{
\item{data}{AlignmentData object.}
}
\value{
Matrix representing the mean across subjects.
}
\description{
Compute the element-wise mean across subjects.
}
