% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment_internal.R
\name{.apply_transforms}
\alias{.apply_transforms}
\title{Internal: Apply Transforms to Data}
\usage{
.apply_transforms(model, data)
}
\description{
Internal: Apply Transforms to Data
}
\keyword{internal}
