% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{.compute_correlation_metrics}
\alias{.compute_correlation_metrics}
\title{Compute Correlation Metrics}
\usage{
.compute_correlation_metrics(aligned)
}
\description{
Compute Correlation Metrics
}
\keyword{internal}
