% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{.compute_distance_matrix}
\alias{.compute_distance_matrix}
\title{Compute Distance Matrix Between Subjects}
\usage{
.compute_distance_matrix(data, distance)
}
\arguments{
\item{data}{AlignmentData object.}

\item{distance}{Distance metric.}
}
\value{
Square distance matrix.
}
\description{
Compute Distance Matrix Between Subjects
}
\keyword{internal}
