% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_ot.R
\name{.compute_gw_barycenter}
\alias{.compute_gw_barycenter}
\title{Compute GW Barycenter}
\usage{
.compute_gw_barycenter(data, epsilon, max_iter, tol)
}
\description{
Compute GW Barycenter
}
\keyword{internal}
