% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{.compute_reconstruction_metrics}
\alias{.compute_reconstruction_metrics}
\title{Compute Reconstruction Metrics}
\usage{
.compute_reconstruction_metrics(aligned, reference)
}
\description{
Compute Reconstruction Metrics
}
\keyword{internal}
