% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_coupled_diag.R
\docType{data}
\name{.coupled_diag_capabilities}
\alias{.coupled_diag_capabilities}
\title{Coupled Diagonalization Capabilities}
\format{
An object of class \code{list} of length 13.
}
\usage{
.coupled_diag_capabilities
}
\description{
Coupled Diagonalization Capabilities
}
\keyword{internal}
