% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_ot.R
\name{.extract_pair_plan}
\alias{.extract_pair_plan}
\title{Extract a pairwise transport plan from a packed (i,j) list}
\usage{
.extract_pair_plan(plans, i, j, n)
}
\arguments{
\item{plans}{List of plans in packed order.}

\item{i}{First domain index.}

\item{j}{Second domain index.}

\item{n}{Total number of domains.}
}
\value{
Transport plan matrix for the (i,j) pair.
}
\description{
manifoldalign stores pairwise plans in the order:
(1,2), (1,3), ..., (1,n), (2,3), (2,4), ..., (n-1,n)
}
\keyword{internal}
