% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment.R
\name{.fit_cv_folds}
\alias{.fit_cv_folds}
\title{Internal: Generic CV with Provided Folds}
\usage{
.fit_cv_folds(data, aligner, reference, cv_folds, compute_quality, ...)
}
\description{
Internal: Generic CV with Provided Folds
}
\keyword{internal}
