% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_alignment.R
\name{.fit_transform_for_subject}
\alias{.fit_transform_for_subject}
\title{Internal: Fit Transform for a Single Subject}
\usage{
.fit_transform_for_subject(aligner, model, data, subject, ...)
}
\description{
Internal: Fit Transform for a Single Subject
}
\keyword{internal}
