% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_kema.R
\docType{data}
\name{.kema_capabilities}
\alias{.kema_capabilities}
\title{KEMA Capabilities}
\format{
An object of class \code{list} of length 13.
}
\usage{
.kema_capabilities
}
\description{
KEMA Capabilities
}
\keyword{internal}
