% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes.R
\name{.procrustes_fit_manifoldalign}
\alias{.procrustes_fit_manifoldalign}
\title{Procrustes Fit via manifoldalign}
\usage{
.procrustes_fit_manifoldalign(
  data_list,
  reference,
  train_data,
  scale,
  reflection,
  tol,
  max_iter,
  ...
)
}
\description{
Procrustes Fit via manifoldalign
}
\keyword{internal}
