% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes_graph.R
\name{.procrustes_graph_apply}
\alias{.procrustes_graph_apply}
\title{Procrustes Graph Apply Function (New Subjects)}
\usage{
.procrustes_graph_apply(fit_result, new_data, ...)
}
\description{
Procrustes Graph Apply Function (New Subjects)
}
\keyword{internal}
