% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes_graph.R
\docType{data}
\name{.procrustes_graph_capabilities}
\alias{.procrustes_graph_capabilities}
\title{Procrustes Graph Capabilities}
\format{
An object of class \code{list} of length 13.
}
\usage{
.procrustes_graph_capabilities
}
\description{
Procrustes Graph Capabilities
}
\keyword{internal}
