% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes.R
\name{.procrustes_truncated_svd}
\alias{.procrustes_truncated_svd}
\title{Truncated-SVD Procrustes Helper}
\usage{
.procrustes_truncated_svd(X, Y, k)
}
\description{
Truncated-SVD Procrustes Helper
}
\keyword{internal}
