% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_coupled_diag.R
\name{.register_coupled_diag}
\alias{.register_coupled_diag}
\title{Register Coupled Diagonalization Aligner}
\usage{
.register_coupled_diag()
}
\description{
Register Coupled Diagonalization Aligner
}
\keyword{internal}
