% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes.R
\name{.register_procrustes}
\alias{.register_procrustes}
\title{Register Procrustes Aligner}
\usage{
.register_procrustes()
}
\description{
Called during package load to register the Procrustes aligner.
}
\keyword{internal}
