% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes_graph.R
\name{.register_procrustes_graph}
\alias{.register_procrustes_graph}
\title{Register Procrustes Graph Aligner}
\usage{
.register_procrustes_graph()
}
\description{
Register Procrustes Graph Aligner
}
\keyword{internal}
