% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.validate_aligner_requirements}
\alias{.validate_aligner_requirements}
\title{Validate Aligner Requirements}
\usage{
.validate_aligner_requirements(name, data)
}
\arguments{
\item{name}{Aligner name.}

\item{data}{AlignmentData object.}
}
\value{
TRUE if valid, otherwise throws an error.
}
\description{
Check that data meets the requirements of an aligner.
}
\keyword{internal}
