% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{.validate_reference_for_method}
\alias{.validate_reference_for_method}
\title{Validate Reference for Aligner}
\usage{
.validate_reference_for_method(reference, method)
}
\arguments{
\item{reference}{Reference specification.}

\item{method}{Aligner method name.}
}
\value{
TRUE if valid, throws error otherwise.
}
\description{
Check if a reference specification is valid for a given aligner.
}
\keyword{internal}
