% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{get_fold_assignments}
\alias{get_fold_assignments}
\title{Get Fold Assignments from Result}
\usage{
get_fold_assignments(result)
}
\arguments{
\item{result}{An AlignmentResult.}
}
\value{
Named vector of fold assignments, or NULL if not CV.
}
\description{
Get Fold Assignments from Result
}
