% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{get_model}
\alias{get_model}
\title{Get Model from Result}
\usage{
get_model(result)
}
\arguments{
\item{result}{An AlignmentResult object.}
}
\value{
The AlignmentModel.
}
\description{
Get Model from Result
}
