% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{get_reference}
\alias{get_reference}
\title{Get Reference Data from Model}
\usage{
get_reference(model)
}
\arguments{
\item{model}{An AlignmentModel object.}
}
\value{
The reference data (matrix or template object).
}
\description{
Get Reference Data from Model
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
ref <- get_reference(get_model(res))
is.null(ref) || is.matrix(ref)

}
\seealso{
\code{\link{get_reference_spec}} for the reference specification.

Other alignment_model:
\code{\link{AlignmentModel}},
\code{\link{AlignmentModel-class}},
\code{\link{add_transform}()},
\code{\link{get_reference_spec}()},
\code{\link{get_transform}()},
\code{\link{get_transforms}()},
\code{\link{has_transform}()},
\code{\link{model_subjects}()},
\code{\link{spaces_compatible}()}
}
\concept{alignment_model}
