% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_reference_data}
\alias{get_reference_data}
\title{Get Reference Data}
\usage{
get_reference_data(data, reference)
}
\arguments{
\item{data}{AlignmentData object.}

\item{reference}{Subject ID or "consensus".}
}
\value{
Matrix representing the reference data.
}
\description{
Get the data for a reference subject, or compute reference if "consensus".
}
