% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_data.R
\name{get_subject_data}
\alias{get_subject_data}
\title{Get Subject Data by ID}
\usage{
get_subject_data(object, subject)
}
\arguments{
\item{object}{An AlignmentData object.}

\item{subject}{Subject ID (character) or index (integer).}
}
\value{
The data for the specified subject.
}
\description{
Get Subject Data by ID
}
