% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{get_transform}
\alias{get_transform}
\title{Get Transform for a Subject}
\usage{
get_transform(model, subject)
}
\arguments{
\item{model}{An AlignmentModel object.}

\item{subject}{Subject ID.}
}
\value{
The transform operator for the subject.
}
\description{
Get Transform for a Subject
}
