% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{\%*\%,AlignmentModel,matrix-method}
\alias{\%*\%,AlignmentModel,matrix-method}
\title{Compose Transform with Data}
\usage{
\S4method{\%*\%}{AlignmentModel,matrix}(x, y)
}
\arguments{
\item{x}{An AlignmentModel.}

\item{y}{Data (matrix or AlignmentData).}
}
\value{
Transformed data.
}
\description{
Apply an alignment model to data using matrix multiplication syntax.
}
