% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{length,AlignmentModel-method}
\alias{length,AlignmentModel-method}
\title{Get Number of Subjects in Model}
\usage{
\S4method{length}{AlignmentModel}(x)
}
\arguments{
\item{x}{An AlignmentModel object.}
}
\value{
Integer number of subjects with transforms.
}
\description{
Get Number of Subjects in Model
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
length(get_model(res))

}
