% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{length,AlignmentResult-method}
\alias{length,AlignmentResult-method}
\title{Get Number of Aligned Subjects}
\usage{
\S4method{length}{AlignmentResult}(x)
}
\arguments{
\item{x}{An AlignmentResult object.}
}
\value{
Integer number of aligned subjects.
}
\description{
Get Number of Aligned Subjects
}
