% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligned_study.R
\name{observation_data}
\alias{observation_data}
\title{Observation Metadata for a Subject Block}
\usage{
observation_data(x, subject = NULL)
}
\arguments{
\item{x}{An \link{AlignedStudy}.}

\item{subject}{Subject id (required when \code{x} has multiple blocks).}
}
\value{
A \code{data.frame}.
}
\description{
Observation Metadata for a Subject Block
}
\seealso{
Other aligned_study:
\code{\link{AlignedBlock}()},
\code{\link{AlignedResampleSet}},
\code{\link{AlignedResampleSet-class}},
\code{\link{AlignedStudy}},
\code{\link{AlignedStudy-class}},
\code{\link{SharedFeatureSpace}()},
\code{\link{align_study}()},
\code{\link{aligned_matrix}()},
\code{\link{alignment_model}()},
\code{\link{analysis_lineage}()},
\code{\link{analysis_safety}()},
\code{\link{analysis_safety_record}()},
\code{\link{as_aligned_resample_set}()},
\code{\link{as_aligned_study}()},
\code{\link{as_alignment_data.AlignedStudy}()},
\code{\link{as_hyperdesign}()},
\code{\link{as_subject_matrices}()},
\code{\link{assert_analysis_safe}()},
\code{\link{assert_common_shared_space}()},
\code{\link{feature_data}()},
\code{\link{map_subjects}()},
\code{\link{orientation}},
\code{\link{shared_feature_space_from_model}()},
\code{\link{shared_space}()},
\code{\link{shared_space_id}()},
\code{\link{shared_spaces_compatible}()},
\code{\link{stack_aligned_resamples}()},
\code{\link{stack_subjects}()},
\code{\link{subject_data}()},
\code{\link{subject_ids}()}
}
\concept{aligned_study}
