% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_alignment_data}
\alias{preprocess_alignment_data}
\title{Preprocess AlignmentData}
\usage{
preprocess_alignment_data(data, ...)
}
\arguments{
\item{data}{\code{AlignmentData} or a named list of matrices.}

\item{...}{Arguments passed to \code{\link{preprocess_matrix}}.}
}
\value{
An \code{AlignmentData} object with preprocessed subject matrices.
}
\description{
Apply \code{\link{preprocess_matrix}} to each subject matrix in an
\code{\link{AlignmentData}} object (or a list coercible to \code{AlignmentData}).
}
