% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_feature_blocks}
\alias{preprocess_feature_blocks}
\title{Preprocess Feature Blocks}
\usage{
preprocess_feature_blocks(blocks_by_subject, ...)
}
\arguments{
\item{blocks_by_subject}{Named list of subjects, each containing a named list
of \code{"alignment_feature_block"} objects.}

\item{...}{Arguments passed to \code{\link{preprocess_matrix}}.}
}
\value{
A \code{blocks_by_subject} list with the same structure and metadata,
but with preprocessed \code{$x} matrices.
}
\description{
Apply \code{\link{preprocess_matrix}} to the \code{$x} matrix of each
\code{\link{alignment_feature_block}} in a \code{blocks_by_subject} structure.
}
