% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_fmrigds.R
\name{print.neuralign_map_family}
\alias{print.neuralign_map_family}
\title{Print Method for neuralign_map_family}
\usage{
\method{print}{neuralign_map_family}(x, ...)
}
\arguments{
\item{x}{A neuralign_map_family object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns x.
}
\description{
Print Method for neuralign_map_family
}
