% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leakage.R
\name{print_leakage_assessment}
\alias{print_leakage_assessment}
\title{Print Leakage Risk Assessment}
\usage{
print_leakage_assessment(risk)
}
\arguments{
\item{risk}{Risk assessment from assess_leakage_risk.}
}
\description{
Print Leakage Risk Assessment
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(20), 4, 5),
  s2 = matrix(rnorm(20), 4, 5),
  s3 = matrix(rnorm(20), 4, 5)
))
res <- fit_alignment(adat, method = "procrustes", reference = "medoid", cv = "loso",
  compute_quality = FALSE
)
print_leakage_assessment(assess_leakage_risk(res))

}
