% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace.R
\name{restrict_operator_to_subspace}
\alias{restrict_operator_to_subspace}
\title{Restrict an Operator to a Subspace}
\usage{
restrict_operator_to_subspace(Q, basis)
}
\arguments{
\item{Q}{Square \verb{d x d} operator.}

\item{basis}{Orthonormal basis matrix \verb{d x r}.}
}
\value{
r-by-r matrix.
}
\description{
Compute the r-by-r operator representing \code{Q} in the coordinates of \code{basis}:
\deqn{Q_r = B^\top Q B.}
}
\details{
For any \code{z} in \code{R^r}, defining \verb{x = B z} (a vector in the subspace), we have:
\deqn{B^\top (Q x) = Q_r z.}
}
\examples{
set.seed(1)
X <- matrix(rnorm(12), 3, 4)
sb <- identified_subspace_basis(X, convention = "right")
Q <- diag(4)
Qr <- restrict_operator_to_subspace(Q, sb$basis)
dim(Qr)

}
