% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_data.R
\name{show,AlignmentData-method}
\alias{show,AlignmentData-method}
\title{Print Method for AlignmentData}
\usage{
\S4method{show}{AlignmentData}(object)
}
\arguments{
\item{object}{An AlignmentData object.}
}
\description{
Print Method for AlignmentData
}
