% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{show,AlignmentModel-method}
\alias{show,AlignmentModel-method}
\title{Print Method for AlignmentModel}
\usage{
\S4method{show}{AlignmentModel}(object)
}
\arguments{
\item{object}{An AlignmentModel object.}
}
\description{
Print Method for AlignmentModel
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
get_model(res)

}
