% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{show,AlignmentResult-method}
\alias{show,AlignmentResult-method}
\title{Print Method for AlignmentResult}
\usage{
\S4method{show}{AlignmentResult}(object)
}
\arguments{
\item{object}{An AlignmentResult object.}
}
\description{
Print Method for AlignmentResult
}
