% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_data.R
\name{sub-AlignmentData}
\alias{sub-AlignmentData}
\alias{[,AlignmentData,ANY,ANY,ANY-method}
\title{Subset AlignmentData by Subject Index or ID}
\usage{
\S4method{[}{AlignmentData,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{An AlignmentData object.}

\item{i}{Integer indices or character subject IDs.}
}
\value{
A new AlignmentData object containing only the selected subjects.
}
\description{
Subset AlignmentData by Subject Index or ID
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4),
  s3 = matrix(rnorm(12), 3, 4)
))
adat[c("s1", "s3")]

}
