% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{to_analysis_matrix}
\alias{to_analysis_matrix}
\title{Transpose Algorithm Matrix to Analysis Orientation}
\usage{
to_analysis_matrix(x)
}
\arguments{
\item{x}{Matrix in features × observations orientation.}
}
\value{
Matrix in observations × features orientation.
}
\description{
Transpose Algorithm Matrix to Analysis Orientation
}
\keyword{internal}
