% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_infix.R
\name{utils_infix}
\alias{utils_infix}
\alias{\%||\%}
\title{Infix Helpers}
\usage{
a \%||\% b
}
\description{
Internal infix operators used throughout the package.

Returns \code{y} when \code{x} is \code{NULL}, otherwise \code{x}.
}
\keyword{internal}
