% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leakage.R
\name{validate_cv_setup}
\alias{validate_cv_setup}
\title{Validate CV Setup for Leakage}
\usage{
validate_cv_setup(cv_folds, reference = "medoid")
}
\arguments{
\item{cv_folds}{CV fold info from create_cv_folds.}

\item{reference}{Reference selection method or subject.}
}
\value{
TRUE if valid, otherwise throws error.
}
\description{
Validate that a cross-validation setup doesn't have leakage.
}
\seealso{
Other leakage: 
\code{\link{assess_leakage_risk}()},
\code{\link{check_obs_leakage}()},
\code{\link{leakage}}
}
\concept{leakage}
