% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{aligner_capabilities}
\alias{aligner_capabilities}
\title{Get Aligner Capabilities}
\usage{
aligner_capabilities(name)
}
\arguments{
\item{name}{Character string identifying the method.}
}
\value{
List of capability flags, or NULL if method not found.
}
\description{
Get Aligner Capabilities
}
\examples{
aligner_capabilities("procrustes")$transform_type

}
\seealso{
Other registry: 
\code{\link{available_aligners}()},
\code{\link{get_aligner}()},
\code{\link{is_aligner_registered}()},
\code{\link{register_aligner}()},
\code{\link{unregister_aligner}()},
\code{\link{validate_aligner_contract}()}
}
\concept{registry}
