% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_kema.R
\name{aligner_kema}
\alias{aligner_kema}
\title{KEMA Alignment}
\description{
Kernel Manifold Alignment (KEMA) via manifoldalign. Produces per-subject
projection operators into a shared latent space.
}
\keyword{internal}
