% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\docType{class}
\name{AlignmentResult-class}
\alias{AlignmentResult-class}
\title{AlignmentResult Class}
\description{
Represents the result of applying an alignment model to data.
Contains the aligned data along with quality metrics and CV info.
}
\section{Slots}{

\describe{
\item{\code{model}}{The AlignmentModel used to produce this result.}

\item{\code{aligned}}{Named list of aligned data (lazy or realized).}

\item{\code{quality}}{List of quality diagnostics.}

\item{\code{cv_info}}{Cross-validation information (fold assignments, etc.).}
}}

