% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{AlignmentResult}
\alias{AlignmentResult}
\title{Create an AlignmentResult Object}
\usage{
AlignmentResult(model, aligned, quality = list(), cv_info = list())
}
\arguments{
\item{model}{The AlignmentModel used.}

\item{aligned}{Named list of aligned data.}

\item{quality}{Optional list of quality metrics.}

\item{cv_info}{Optional cross-validation information.}
}
\value{
An AlignmentResult object.
}
\description{
Create an AlignmentResult Object
}
\seealso{
Other alignment_result: 
\code{\link{AlignmentResult-class}},
\code{\link{get_aligned}()},
\code{\link{get_cv_info}()},
\code{\link{get_model}()},
\code{\link{get_quality}()}
}
\concept{alignment_result}
