% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_fmrigds.R
\name{apply_to_gds}
\alias{apply_to_gds}
\title{Apply AlignmentModel in fmrigds Pipeline}
\usage{
apply_to_gds(model, gds_data, ...)
}
\arguments{
\item{model}{AlignmentModel.}

\item{gds_data}{A gds_data object from fmrigds.}

\item{...}{Additional arguments.}
}
\value{
Transformed gds_data object.
}
\description{
Helper for applying neuralign alignment within fmrigds workflows.
}
