% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_alignment.R
\name{apply_transform}
\alias{apply_transform}
\title{Apply Transform to Data}
\usage{
apply_transform(transform, data)
}
\arguments{
\item{transform}{The transform operator (matrix, target x source).}

\item{data}{Data to transform (matrix, features x observations).}
}
\value{
Transformed data.
}
\description{
Apply a single transform operator to data using left-multiply convention.
}
