% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_data.R
\name{as_alignment_data}
\alias{as_alignment_data}
\title{Coerce List to AlignmentData}
\usage{
as_alignment_data(x, ...)
}
\arguments{
\item{x}{A named list of matrices.}

\item{...}{Additional arguments passed to AlignmentData constructor.}
}
\value{
An AlignmentData object.
}
\description{
Coerce List to AlignmentData
}
