% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment.R
\name{.compute_basic_quality}
\alias{.compute_basic_quality}
\title{Internal: Compute Basic Quality Metrics}
\usage{
.compute_basic_quality(data, aligned, model)
}
\description{
Internal: Compute Basic Quality Metrics
}
\keyword{internal}
