% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{.compute_pairwise_distance}
\alias{.compute_pairwise_distance}
\title{Compute Distance Between Two Matrices}
\usage{
.compute_pairwise_distance(x, y, distance)
}
\arguments{
\item{x}{First matrix.}

\item{y}{Second matrix.}

\item{distance}{Distance metric.}
}
\value{
Scalar distance.
}
\description{
Compute Distance Between Two Matrices
}
\keyword{internal}
