% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment.R
\name{.fit_cv_kfold}
\alias{.fit_cv_kfold}
\title{Internal: K-Fold CV}
\usage{
.fit_cv_kfold(
  data,
  aligner,
  reference,
  k,
  compute_quality,
  return_aligned = TRUE,
  ...
)
}
\description{
Internal: K-Fold CV
}
\keyword{internal}
