% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment_internal.R
\name{.fit_cv_loso}
\alias{.fit_cv_loso}
\title{Internal: Leave-One-Subject-Out CV}
\usage{
.fit_cv_loso(
  data,
  aligner,
  reference,
  compute_quality,
  return_aligned = TRUE,
  return_resample_artifacts = FALSE,
  restrict_to_identified = FALSE,
  restrict_tol = sqrt(.Machine$double.eps),
  ...
)
}
\description{
Internal: Leave-One-Subject-Out CV
}
\keyword{internal}
