% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment_internal.R
\name{.fit_new_subject}
\alias{.fit_new_subject}
\title{Internal: Fit Transform for New Subject}
\usage{
.fit_new_subject(
  aligner,
  fit_result,
  data,
  subject_idx,
  reference,
  fold_id = NULL
)
}
\description{
Internal: Fit Transform for New Subject
}
\keyword{internal}
