% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment.R
\name{.fit_single}
\alias{.fit_single}
\title{Internal: Single Fit (No CV)}
\usage{
.fit_single(
  data,
  aligner,
  reference,
  train_idx,
  compute_quality,
  return_aligned = TRUE,
  restrict_to_identified = FALSE,
  restrict_tol = sqrt(.Machine$double.eps),
  ...
)
}
\description{
Internal: Single Fit (No CV)
}
\keyword{internal}
