% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_ot.R
\docType{data}
\name{.fpgw_capabilities}
\alias{.fpgw_capabilities}
\title{FPGW Capabilities}
\format{
An object of class \code{list} of length 13.
}
\usage{
.fpgw_capabilities
}
\description{
FPGW Capabilities
}
\keyword{internal}
