% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_ot.R
\name{.fpgw_fit}
\alias{.fpgw_fit}
\title{Fused Procrustes-GW Fit Function}
\usage{
.fpgw_fit(
  data,
  reference = "medoid",
  train_idx = NULL,
  omega1 = 0.001,
  epsilon = 0.01,
  max_iter = 100,
  tol = 1e-06,
  fit_context = NULL,
  provider_plan = NULL,
  ...
)
}
\description{
Combines Procrustes rotation with GW optimal transport.
}
\keyword{internal}
