% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes.R
\name{.gpa_builtin}
\alias{.gpa_builtin}
\title{Built-in GPA Implementation}
\usage{
.gpa_builtin(data_list, scale, reflection, rank, tol, max_iter)
}
\description{
Built-in GPA Implementation
}
\keyword{internal}
