% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_gpca.R
\name{.gpca_fit}
\alias{.gpca_fit}
\title{GPCA Fit Function}
\usage{
.gpca_fit(
  data,
  reference = "medoid",
  train_idx = NULL,
  ncomp = 10L,
  u = 0.5,
  lambda = 0.01,
  ...
)
}
\arguments{
\item{data}{AlignmentData object.}

\item{reference}{Reference specification.}

\item{train_idx}{Indices of subjects to use for fitting.}

\item{ncomp}{Number of components.}

\item{u}{Trade-off parameter (0-1).}

\item{lambda}{Regularisation.}

\item{...}{Additional arguments forwarded to manifoldalign::gpca_align().}
}
\value{
List with transforms, reference_data, etc.
}
\description{
GPCA Fit Function
}
\keyword{internal}
