% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_grasp.R
\name{.grasp_fit}
\alias{.grasp_fit}
\title{GRASP Fit Function}
\usage{
.grasp_fit(
  data,
  reference = "medoid",
  train_idx = NULL,
  ncomp = 30L,
  q_descriptors = 100L,
  sigma = 0.73,
  lambda = 0.1,
  solver = "linear",
  ...
)
}
\arguments{
\item{data}{AlignmentData object.}

\item{reference}{Reference specification.}

\item{train_idx}{Indices of subjects to use for fitting.}

\item{ncomp}{Number of spectral components.}

\item{q_descriptors}{Number of diffusion-time descriptors.}

\item{sigma}{Diffusion bandwidth.}

\item{lambda}{Regularisation for base alignment.}

\item{solver}{Assignment solver: "linear" or "hungarian".}

\item{...}{Additional arguments forwarded to manifoldalign::grasp()
or manifoldalign::grasp_multiset().}
}
\value{
List with transforms, reference_data, etc.
}
\description{
GRASP Fit Function
}
\keyword{internal}
