% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_ot.R
\docType{data}
\name{.gw_capabilities}
\alias{.gw_capabilities}
\title{GW Capabilities}
\format{
An object of class \code{list} of length 11.
}
\usage{
.gw_capabilities
}
\description{
GW Capabilities
}
\keyword{internal}
