% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes.R
\docType{data}
\name{.procrustes_capabilities}
\alias{.procrustes_capabilities}
\title{Procrustes Capabilities}
\format{
An object of class \code{list} of length 13.
}
\usage{
.procrustes_capabilities
}
\description{
Procrustes Capabilities
}
\keyword{internal}
