% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes.R
\name{.procrustes_fit_builtin}
\alias{.procrustes_fit_builtin}
\title{Built-in Procrustes Implementation}
\usage{
.procrustes_fit_builtin(
  data_list,
  reference,
  train_data,
  scale,
  reflection,
  tol,
  max_iter
)
}
\description{
Built-in Procrustes Implementation
}
\keyword{internal}
