% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes_graph.R
\name{.procrustes_graph_fit}
\alias{.procrustes_graph_fit}
\title{Procrustes Graph Fit Function}
\usage{
.procrustes_graph_fit(
  data,
  reference = NULL,
  train_idx = NULL,
  min_overlap = 2L,
  weight = c("overlap", "uniform"),
  reflection = FALSE,
  allow_reflection = NULL,
  fit_context = NULL,
  provider_plan = NULL,
  ...
)
}
\description{
Procrustes Graph Fit Function
}
\keyword{internal}
