% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner_procrustes.R
\name{.procrustes_single}
\alias{.procrustes_single}
\title{Single Procrustes Alignment}
\usage{
.procrustes_single(X, Y, scale = FALSE, reflection = FALSE, rank = NULL)
}
\description{
Single Procrustes Alignment
}
\keyword{internal}
