% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment.R
\name{.resolve_reference_spec}
\alias{.resolve_reference_spec}
\title{Internal: Resolve Reference Specification}
\usage{
.resolve_reference_spec(data, reference, train_idx)
}
\description{
Internal: Resolve Reference Specification
}
\keyword{internal}
