% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment_internal.R
\name{.resolve_unevaluated_subjects}
\alias{.resolve_unevaluated_subjects}
\title{Internal: Generic CV with Provided Folds}
\usage{
.resolve_unevaluated_subjects(cv_folds, subjects)
}
\description{
Internal: Generic CV with Provided Folds
}
\keyword{internal}
