% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alignment_internal.R
\name{.validate_cv_folds_spec}
\alias{.validate_cv_folds_spec}
\title{Internal: Validate Fold Spec}
\usage{
.validate_cv_folds_spec(cv_folds, n_subjects)
}
\description{
Internal: Validate Fold Spec
}
\keyword{internal}
