% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_fmrigds.R
\name{from_map_family}
\alias{from_map_family}
\title{Convert MapFamily to AlignmentModel}
\usage{
from_map_family(map_family, method = "fmrigds_imported")
}
\arguments{
\item{map_family}{A MapFamily object from fmrigds.}

\item{method}{Optional method name to assign.}
}
\value{
An AlignmentModel.
}
\description{
Convert an fmrigds MapFamily to a neuralign AlignmentModel.
}
