% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{get_aligned}
\alias{get_aligned}
\title{Get Aligned Data for a Subject}
\usage{
get_aligned(result, subject = NULL)
}
\arguments{
\item{result}{An AlignmentResult object.}

\item{subject}{Subject ID.}
}
\value{
The aligned data for the subject.
}
\description{
Get Aligned Data for a Subject
}
