% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_aligner}
\alias{get_aligner}
\title{Get Information About an Aligner}
\usage{
get_aligner(name)
}
\arguments{
\item{name}{Character string identifying the method.}
}
\value{
List with aligner information, or NULL if not found.
}
\description{
Get Information About an Aligner
}
\examples{
get_aligner("procrustes")$description

}
\seealso{
Other registry:
\code{\link{aligner_capabilities}()},
\code{\link{available_aligners}()},
\code{\link{is_aligner_registered}()},
\code{\link{register_aligner}()},
\code{\link{unregister_aligner}()},
\code{\link{validate_aligner_contract}()}
}
\concept{registry}
