% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{get_cv_info}
\alias{get_cv_info}
\title{Get CV Info from Result}
\usage{
get_cv_info(result)
}
\arguments{
\item{result}{An AlignmentResult object.}
}
\value{
List of CV information.
}
\description{
Get CV Info from Result
}
