% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_data.R
\name{get_data_list}
\alias{get_data_list}
\title{Get All Data as a List}
\usage{
get_data_list(object)
}
\arguments{
\item{object}{An AlignmentData object.}
}
\value{
Named list of subject data.
}
\description{
Get All Data as a List
}
