% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_data.R
\name{get_obs_labels}
\alias{get_obs_labels}
\title{Get Observation Labels}
\usage{
get_obs_labels(object)
}
\arguments{
\item{object}{An AlignmentData object.}
}
\value{
Observation labels (or NULL if not set).
}
\description{
Get Observation Labels
}
