% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{get_quality}
\alias{get_quality}
\title{Get Quality Metrics}
\usage{
get_quality(result, metric = NULL)
}
\arguments{
\item{result}{An AlignmentResult object.}

\item{metric}{Optional specific metric name.}
}
\value{
Quality metrics list or specific metric value.
}
\description{
Get Quality Metrics
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(20), 4, 5),
  s2 = matrix(rnorm(20), 4, 5)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = TRUE)
qm <- get_quality(res)
names(qm)

}
\seealso{
Other alignment_result: 
\code{\link{AlignmentResult}},
\code{\link{AlignmentResult-class}},
\code{\link{get_aligned}()},
\code{\link{get_cv_info}()},
\code{\link{get_model}()}
}
\concept{alignment_result}
