% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{get_reference}
\alias{get_reference}
\title{Get Reference Data from Model}
\usage{
get_reference(model)
}
\arguments{
\item{model}{An AlignmentModel object.}
}
\value{
The reference data (matrix or template object).
}
\description{
Get Reference Data from Model
}
\seealso{
\code{\link{get_reference_spec}} for the reference specification.
}
