% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_data.R
\name{get_subject_data}
\alias{get_subject_data}
\title{Get Subject Data by ID}
\usage{
get_subject_data(object, subject)
}
\arguments{
\item{object}{An AlignmentData object.}

\item{subject}{Subject ID (character) or index (integer).}
}
\value{
The data for the specified subject.
}
\description{
Get Subject Data by ID
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4)
))
x1 <- get_subject_data(adat, "s1")
dim(x1)

}
\seealso{
Other alignment_data:
\code{\link{AlignmentData}},
\code{\link{AlignmentData-class}},
\code{\link{as_alignment_data}()},
\code{\link{get_data_list}()},
\code{\link{get_obs_labels}()},
\code{\link{validate_alignment_data}()}
}
\concept{alignment_data}
