% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{get_transform}
\alias{get_transform}
\title{Get Transform for a Subject}
\usage{
get_transform(model, subject)
}
\arguments{
\item{model}{An AlignmentModel object.}

\item{subject}{Subject ID.}
}
\value{
The transform operator for the subject.
}
\description{
Get Transform for a Subject
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
mdl <- get_model(res)
Q_s2 <- get_transform(mdl, "s2")
dim(Q_s2)

}
\seealso{
Other alignment_model: 
\code{\link{AlignmentModel}},
\code{\link{AlignmentModel-class}},
\code{\link{add_transform}()},
\code{\link{get_reference}()},
\code{\link{get_reference_spec}()},
\code{\link{get_transforms}()},
\code{\link{has_transform}()},
\code{\link{model_subjects}()},
\code{\link{spaces_compatible}()}
}
\concept{alignment_model}
