% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{get_transforms}
\alias{get_transforms}
\title{Get All Transforms}
\usage{
get_transforms(model)
}
\arguments{
\item{model}{An AlignmentModel object.}
}
\value{
Named list of transform operators.
}
\description{
Get All Transforms
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
trs <- get_transforms(get_model(res))
names(trs)

}
\seealso{
Other alignment_model:
\code{\link{AlignmentModel}},
\code{\link{AlignmentModel-class}},
\code{\link{add_transform}()},
\code{\link{get_reference}()},
\code{\link{get_reference_spec}()},
\code{\link{get_transform}()},
\code{\link{has_transform}()},
\code{\link{model_subjects}()},
\code{\link{spaces_compatible}()}
}
\concept{alignment_model}
