% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{get_transforms}
\alias{get_transforms}
\title{Get All Transforms}
\usage{
get_transforms(model)
}
\arguments{
\item{model}{An AlignmentModel object.}
}
\value{
Named list of transform operators.
}
\description{
Get All Transforms
}
