% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{has_transform}
\alias{has_transform}
\title{Check if Model Has Transform for Subject}
\usage{
has_transform(model, subject)
}
\arguments{
\item{model}{An AlignmentModel object.}

\item{subject}{Subject ID.}
}
\value{
Logical.
}
\description{
Check if Model Has Transform for Subject
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
mdl <- get_model(res)
has_transform(mdl, "s2")
has_transform(mdl, "missing")

}
\seealso{
Other alignment_model:
\code{\link{AlignmentModel}},
\code{\link{AlignmentModel-class}},
\code{\link{add_transform}()},
\code{\link{get_reference}()},
\code{\link{get_reference_spec}()},
\code{\link{get_transform}()},
\code{\link{get_transforms}()},
\code{\link{model_subjects}()},
\code{\link{spaces_compatible}()}
}
\concept{alignment_model}
