% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{has_transform}
\alias{has_transform}
\title{Check if Model Has Transform for Subject}
\usage{
has_transform(model, subject)
}
\arguments{
\item{model}{An AlignmentModel object.}

\item{subject}{Subject ID.}
}
\value{
Logical.
}
\description{
Check if Model Has Transform for Subject
}
