% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{is_aligner_registered}
\alias{is_aligner_registered}
\title{Check if Aligner is Registered}
\usage{
is_aligner_registered(name)
}
\arguments{
\item{name}{Character string identifying the method.}
}
\value{
Logical indicating if the aligner is registered.
}
\description{
Check if Aligner is Registered
}
