% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{is_cv_result}
\alias{is_cv_result}
\title{Check if Result is Cross-Validated}
\usage{
is_cv_result(result)
}
\arguments{
\item{result}{An AlignmentResult.}
}
\value{
Logical indicating if result used CV.
}
\description{
Check if Result is Cross-Validated
}
