% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_data.R
\name{length,AlignmentData-method}
\alias{length,AlignmentData-method}
\title{Get Number of Subjects in AlignmentData}
\usage{
\S4method{length}{AlignmentData}(x)
}
\arguments{
\item{x}{An AlignmentData object.}
}
\value{
Integer number of subjects.
}
\description{
Get Number of Subjects in AlignmentData
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4)
))
length(adat)

}
