% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{length,AlignmentResult-method}
\alias{length,AlignmentResult-method}
\title{Get Number of Aligned Subjects}
\usage{
\S4method{length}{AlignmentResult}(x)
}
\arguments{
\item{x}{An AlignmentResult object.}
}
\value{
Integer number of aligned subjects.
}
\description{
Get Number of Aligned Subjects
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(20), 4, 5),
  s2 = matrix(rnorm(20), 4, 5)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
length(res)

}
