% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{model_subjects}
\alias{model_subjects}
\title{Get Subjects in Model}
\usage{
model_subjects(model)
}
\arguments{
\item{model}{An AlignmentModel object.}
}
\value{
Character vector of subject IDs.
}
\description{
Get Subjects in Model
}
