% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{NEURALIGN_ALIGNER_API_VERSION}
\alias{NEURALIGN_ALIGNER_API_VERSION}
\title{Aligner API Version}
\format{
An object of class \code{integer} of length 1.
}
\usage{
NEURALIGN_ALIGNER_API_VERSION
}
\description{
The sole supported version of the aligner registration API. Third-party
packages should declare this version so incompatible provider contracts fail
during registration rather than during fitting.
}
\examples{
NEURALIGN_ALIGNER_API_VERSION

}
\keyword{datasets}
