% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leakage.R
\name{print_leakage_assessment}
\alias{print_leakage_assessment}
\title{Print Leakage Risk Assessment}
\usage{
print_leakage_assessment(risk)
}
\arguments{
\item{risk}{Risk assessment from assess_leakage_risk.}
}
\description{
Print Leakage Risk Assessment
}
