% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{show,AlignmentModel-method}
\alias{show,AlignmentModel-method}
\title{Print Method for AlignmentModel}
\usage{
\S4method{show}{AlignmentModel}(object)
}
\arguments{
\item{object}{An AlignmentModel object.}
}
\description{
Print Method for AlignmentModel
}
