% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{show,AlignmentResult-method}
\alias{show,AlignmentResult-method}
\title{Print Method for AlignmentResult}
\usage{
\S4method{show}{AlignmentResult}(object)
}
\arguments{
\item{object}{An AlignmentResult object.}
}
\description{
Print Method for AlignmentResult
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(20), 4, 5),
  s2 = matrix(rnorm(20), 4, 5)
))
fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)

}
