% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{spaces_compatible}
\alias{spaces_compatible}
\title{Check Space Compatibility}
\usage{
spaces_compatible(a, b)
}
\arguments{
\item{a}{First space specification.}

\item{b}{Second space specification.}
}
\value{
Logical; TRUE if spaces are compatible.
}
\description{
Determine whether two space specifications are compatible.
NULL is compatible with anything.
}
\seealso{
Other alignment_model: 
\code{\link{AlignmentModel}},
\code{\link{AlignmentModel-class}},
\code{\link{add_transform}()},
\code{\link{get_reference}()},
\code{\link{get_reference_spec}()},
\code{\link{get_transform}()},
\code{\link{get_transforms}()},
\code{\link{has_transform}()},
\code{\link{model_subjects}()}
}
\concept{alignment_model}
