% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_model.R
\name{sub-AlignmentModel}
\alias{sub-AlignmentModel}
\alias{[,AlignmentModel,ANY,ANY,ANY-method}
\title{Subset AlignmentModel by Subject}
\usage{
\S4method{[}{AlignmentModel,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{An AlignmentModel object.}

\item{i}{Character subject IDs or integer indices.}
}
\value{
A new AlignmentModel with only the selected subjects.
}
\description{
Subset AlignmentModel by Subject
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(12), 3, 4),
  s2 = matrix(rnorm(12), 3, 4),
  s3 = matrix(rnorm(12), 3, 4)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
mdl <- get_model(res)
mdl[c("s1", "s3")]

}
