% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_result.R
\name{sub-AlignmentResult}
\alias{sub-AlignmentResult}
\alias{[,AlignmentResult,ANY,ANY,ANY-method}
\title{Subset AlignmentResult by Subject}
\usage{
\S4method{[}{AlignmentResult,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{An AlignmentResult object.}

\item{i}{Subject IDs or indices.}
}
\value{
A new AlignmentResult with only selected subjects.
}
\description{
Subset AlignmentResult by Subject
}
\examples{
set.seed(1)
adat <- AlignmentData(list(
  s1 = matrix(rnorm(20), 4, 5),
  s2 = matrix(rnorm(20), 4, 5),
  s3 = matrix(rnorm(20), 4, 5)
))
res <- fit_alignment(adat, method = "procrustes", reference = "s1", compute_quality = FALSE)
res[c("s1", "s3")]

}
